requirements.txt
keyboard
numpys